+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## [Unreleased]
 ### Added
 - Making Universal Builds for MacOS.
+- Play Background Music
 
 ### Fixed
 - Making Windows/MacOS/Linux executables automatically per new version.
@@ -18,4 +19,5 @@
 - Player Tracking Camera (Basic)
 - Character Movement via WASD
 
+[Unreleased]: https://github.com/Exonoesis/Helping-Hand/blob/main/CHANGELOG.md
 [0.1.0]: https://github.com/Exonoesis/Helping-Hand/releases/tag/v0.1.0